// packages/medusa-client/src/admin/models/AdminPriceListsProductsListRes.ts
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

import type { Product } from './Product';

export type AdminPriceListsProductsListRes = {
  products?: Array<Product>;
  /**
   * The total number of items available
   */
  count?: number;
  /**
   * The number of items skipped before these items
   */
  offset?: number;
  /**
   * The number of items per page
   */
  limit?: number;
};